added fix for BERTLVAsciiHexPackager getUninterpretLength bug

+++ jpos/src/main/java/org/jpos/tlv/packager/bertlv/BERTLVPackager.java
@@ -447,7 +447,7 @@ public abstract class BERTLVPackager extends GenericPackager {
     private int getUninterpretLength(int length, BinaryInterpreter interpreter) {
         if (length > 0) {
             int lengthAdjusted = length + length % 2;
-            return length * (lengthAdjusted / interpreter.getPackedLength(lengthAdjusted));
+            return (int) (length * (lengthAdjusted / (double) interpreter.getPackedLength(lengthAdjusted)));
         }
         return 0;
     }
